third_party/blink/web_tests/virtual/do-not-hide-visited-color/external/wpt/css/css-pseudo/highlight-cascade/highlight-currentcolor-computed-visited-expected.txt

This is a testharness.js-based test.
[FAIL] getComputedStyle() for ::selection at #target1
  assert_equals: Color is lime. expected "rgb(0, 255, 0)" but got "rgb(0, 255, 255)"
[FAIL] getComputedStyle() for ::target-text at #target1
  assert_equals: Color is lime. expected "rgb(0, 255, 0)" but got "rgb(0, 255, 255)"
[FAIL] getComputedStyle() for ::search-text at #target1
  assert_equals: Color is lime. expected "rgb(0, 255, 0)" but got "rgb(0, 255, 255)"
[FAIL] getComputedStyle() for ::spelling-error at #target1
  assert_equals: Color is lime. expected "rgb(0, 255, 0)" but got "rgb(0, 255, 255)"
[FAIL] getComputedStyle() for ::grammar-error at #target1
  assert_equals: Color is lime. expected "rgb(0, 255, 0)" but got "rgb(0, 255, 255)"
[FAIL] getComputedStyle() for ::highlight(foo) at #target1
  assert_equals: Color is lime. expected "rgb(0, 255, 0)" but got "rgb(0, 255, 255)"
Harness: the test ran to completion.
